Merge from dotnet/runtime

+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Runtime/RuntimeImports.cs
@@ -265,7 +265,7 @@ namespace System.Runtime
             // The runtime performs additional checks in debug builds
             return _RhHandleGet(handle);
 #else
-            return Unsafe.As<IntPtr, object>(ref *(IntPtr*)handle);
+            return Unsafe.As<IntPtr, object>(ref *(IntPtr*)(nint)handle);
 #endif
         }
